Refactor imports for consistency: consolidate slash command imports across multiple files
