+++ CityOfBinds/src/game/utils/__init__.py
@@ -4,6 +4,13 @@ from .powers.power import _Power
 from .triggers.trigger import _Trigger
 from .triggers.wasd_trigger import _WASDTrigger
 from .triggers.trigger_mixin import _TriggerMixin
+from .slash_commands.slash_command import SlashCommand
+from .slash_commands.command_group.command_group import (
+    _CommandGroup,
+    CommandGroupConstants,
+)
+from .slash_commands.command_group.commands_mixin import _CommandsMixin
+
 
 __all__ = [
     "_Comment",
@@ -12,4 +19,8 @@ __all__ = [
     "_Trigger",
     "_WASDTrigger",
     "_TriggerMixin",
+    "SlashCommand",
+    "_CommandGroup",
+    "CommandGroupConstants",
+    "_CommandsMixin",
 ]
