--- CityOfBinds/src/game/utils/slash_commands/__init__.py
@@ -1,4 +1,0 @@
-from .command_group import _CommandGroup, _CommandsMixin, CommandGroupConstants
-from .slash_command import SlashCommand
-
-__all__ = ["SlashCommand", "_CommandGroup", "_CommandsMixin", "CommandGroupConstants"]
