--- CityOfBinds/src/game/utils/slash_commands/command_group/__init__.py
@@ -1,4 +1,0 @@
-from .command_group import _CommandGroup, CommandGroupConstants
-from .command_group_mixin import _CommandsMixin
-
-__all__ = ["_CommandGroup", "_CommandsMixin", "CommandGroupConstants"]
